# [Torch] More graph rewrites for Faster RCNN / MaskRCNN