@masahi 2021-01-27
No, the `If` there is for guarding against the case where there is no boxes, see 

https://github.com/pytorch/vision/blob/8ebfd2f5d5f1792ce2cf5a2329320f604530a68e/torchvision/ops/boxes.py#L78-L79

So applying topk to an empty tensor is nop anyway.